Rename reactGA and including it like the cool kids.
