+++ src/features/app/App.js
@@ -4,7 +4,7 @@ import {
   Row
 } from 'react-bootstrap'
 import { Provider } from 'react-redux'
-import './ReactGA.js'
+import './reactGA'
 import Header from './Header'
 import InputTabs from './InputTabs'
 import MachinePreview from '../machine/MachinePreview'
